+++ src/services/absences/index.js
@@ -6,15 +6,23 @@ export const list = () =>
 
 /* startDate and endDate have to be passed as moment objects,
    to display absences in a calendar view
-   canMove: false => to not move this event in a calendar
 */
 export const listWithMoment = () =>
   list().map((absence) => {
     return {
       ...absence,
       startDate: moment(absence.startDate),
-      endDate: moment(absence.endDate).add(24, 'hour'),
-      canMove: false
+      endDate: moment(absence.endDate).add(24, 'hour')
+    };
+  });
+
+export const listWithMembers = (members) =>
+  listWithMoment().map((event) => {
+    const memberObject = members.filter((member) => member.userId === event.userId)
+      || null;
+    return {
+      ...event,
+      user: memberObject,
     };
   });
 
@@ -25,5 +33,6 @@ export const get = (id) =>
 export default {
   get,
   list,
-  listWithMoment
+  listWithMoment,
+  listWithMembers
 };