+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## CHANGELOG.md
 
+## 1.3.0
+* Generate .ica file to download
+#####TODO:
+  * [ ] FIX: Lines not delimited by CRLF sequence
+  * [ ] test iCalendar service
+  * [ ] test listWithMembers function in absence service
+
 ## 1.2.0
 * Add member's absences in a calendar view
 * Add listWithMoment() function to absence service